Changed Video dto length from int to double to handle cases where length was not returned in whole seconds.
Added paging to vod list, fixed "CanNext" logic for top twitch streams and vod list.

Changed formatting of duration of vods.

+++ TwitchTv/Dto/Video.cs
@@ -27,7 +27,7 @@ namespace TwitchTv.Dto
         public string Game { get; set; }
 
         /// <summary>  Length of the broadcast in seconds </summary>
-        public int Length { get; set; }
+        public double Length { get; set; }
 
         [JsonProperty("delete_at")]
         public DateTime? DeleteAt { get; set; }
